move auht middleware into server package
