+++ server/http/middleware/testdata/response_500.json
